requirements.txt
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.2
defusedxml==0.8.0rc2
Django==5.0.4
django-filter==24.2
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
Markdown==3.6
oauthlib==3.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==2.0.0
service-identity==24.1.0
six==1.16.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.11.0
urllib3==2.2.1
zope.interface==6.3


Temp@1337++!


user6
eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ0b2tlbl90eXBlIjoiYWNjZXNzIiwiZXhwIjoxNzE0NjgyNDA3LCJpYXQiOjE3MTQ1OTYwMDcsImp0aSI6IjliNzcxMDE4OTIyNjQ5N2ZhODY0ZDVlYmY2NjkzODY2IiwidXNlcl9pZCI6InVzZXI2QG1haWwuY29tIn0.gnILT_uuLV2yWP0NLPyd1cBRZBCqMeNmNd1KDKmxpIE

user7

eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ0b2tlbl90eXBlIjoiYWNjZXNzIiwiZXhwIjoxNzE0NjgyOTA2LCJpYXQiOjE3MTQ1OTY1MDYsImp0aSI6ImRjYTBlYTk3MjI2NzQzOTk4ZWRkMGI5OTQ4NGY3MDhiIiwidXNlcl9pZCI6InVzZXI3QGdtYWlsLmNvbSJ9.jNIx-eeECB6FgVreP2wWnV9AR-VAR4iBVhaOyyfj_ps